requirements.txt
llama-cpp-python
numpy
flask
flask-cors
torch
#transformer
